fix(search): bigint token_count + named-vector e2e test_50
- squawk prefer-bigint-over-int on the new chunks.token_count column
- test_50 raw-Qdrant assertions updated to the named-vector shape
  (vectors.dense.size + using:dense); the app /api/search e2e
  (test_search_via_api) already passes against real named-vector Qdrant

Refs #595

+++ priv/repo/migrations/20260615120000_add_token_count_to_chunks.exs
@@ -4,7 +4,7 @@ defmodule Engram.Repo.Migrations.AddTokenCountToChunks do
   # phase/expand — additive nullable column; no backfill (pre-launch, wipeable).
   def change do
     alter table(:chunks) do
-      add :token_count, :integer
+      add :token_count, :bigint
     end
   end
 end
